WASM: Add JWT policy
This is a dirty commit POC around implementation of a policy, there are a lot
of hardcoded things, and these are the reason:

-> On a new filter on rust, no way to share metadata. So config should
be hardcoded, PR to change that in upstream is already open.
-> On JWT parsing, there is no way to get the value from Envoy, shared
metadata, etc.. so need to work on decode that, all the libraries are
using SSL libs, and WASM cannot be started if enabled.

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>